Add weapon and some minor balance and stats adjusting

+++ Archer Evolve/Assets/Scripts/Entity/Entity.cs
@@ -6,10 +6,10 @@ public interface IEntity
 {
     float moveSpeed {  get; set; }
     int healthPoint {  get; set; }
-    int strength {  get; set; }
+    float strength {  get; set; }
     float attackSpeed {  get; set; }
 
     void Moving(Vector2 moveInput);
     void Attack();
-    void TakeDamge(int strength);
+    void TakeDamge(float strength);
 }
